distribute tests by library and rename (#209)
