+++ Tests/SwiftLayoutTests/ImplementationTests.swift
@@ -234,117 +234,6 @@ extension ImplementationTests {
         let constraintsBetweebViews = Set(AnchorsContainer(Anchors.top.equalTo(label!, attribute: .bottom)).constraints(item: secondView!, toItem: label).weakens)
         XCTAssertEqual(currents.intersection(constraintsBetweebViews), constraintsBetweebViews)
     }
-    
-    func testDebugLayoutStructurePrint() {
-        let root: UIView = UIView().identifying("root")
-        let contentView: UIView = UIView().identifying("contentView")
-        let image: UIImageView = UIImageView().identifying("image")
-        let title: UILabel = UILabel().identifying("title")
-        let description: UILabel = UILabel().identifying("description")
-        
-        let layout = root {
-            contentView.anchors {
-                Anchors.leading.equalTo(root.safeAreaLayoutGuide, constant: 16.0)
-                Anchors.trailing.equalTo(root.safeAreaLayoutGuide, constant: -16.0)
-                Anchors.centerY.equalTo(root)
-                Anchors.height.equalTo(constant: 80.0)
-            }.sublayout {
-                image.anchors {
-                    Anchors.leading.equalToSuper(constant: 10.0)
-                    Anchors.centerY
-                    Anchors.size(width: 70, height: 70)
-                }
-                
-                title.anchors {
-                    Anchors.top.equalToSuper(constant: 8.0)
-                    Anchors.leading.equalTo(image, attribute: .trailing, constant: 10.0)
-                    Anchors.height.equalTo(constant: 24.0)
-                }.anyLayout
-                
-                description.anchors {
-                    Anchors.top.equalTo(title, attribute: .bottom, constant: 5.0)
-                    Anchors.leading.equalTo(image, attribute: .trailing, constant: 10.0)
-                    Anchors.bottom.equalToSuper(constant: -8.0)
-                    Anchors.trailing.equalToSuper(constant: -10.0)
-                }
-            }
-        }
-        
-        let expectedResult = """
-        ViewLayout - view: root
-        └─ ViewLayout - view: contentView
-           └─ TupleLayout
-              ├─ ViewLayout - view: image
-              ├─ AnyLayout
-              │  └─ ViewLayout - view: title
-              └─ ViewLayout - view: description
-        """
-        
-        XCTAssertEqual(LayoutPrinter(layout).description, expectedResult)
-    }
-    
-    func testDebugLayoutStructurePrintWithAnchors() {
-        let root: UIView = UIView().identifying("root")
-        let contentView: UIView = UIView().identifying("contentView")
-        let image: UIImageView = UIImageView().identifying("image")
-        let title: UILabel = UILabel().identifying("title")
-        let description: UILabel = UILabel().identifying("description")
-        
-        let layout = root {
-            contentView.anchors {
-                Anchors.leading.equalTo(root.safeAreaLayoutGuide, constant: 16.0)
-                Anchors.trailing.equalTo(root.safeAreaLayoutGuide, constant: -16.0)
-                Anchors.centerY.equalTo(root)
-                Anchors.height.equalTo(constant: 80.0)
-            }.sublayout {
-                image.anchors {
-                    Anchors.leading.equalToSuper(constant: 10.0)
-                    Anchors.centerY
-                    Anchors.size(width: 70, height: 70)
-                }
-                
-                title.anchors {
-                    Anchors.top.equalToSuper(constant: 8.0)
-                    Anchors.leading.equalTo(image, attribute: .trailing, constant: 10.0)
-                    Anchors.height.equalTo(constant: 24.0)
-                }.anyLayout
-                
-                description.anchors {
-                    Anchors.top.equalTo(title, attribute: .bottom, constant: 5.0)
-                    Anchors.leading.equalTo(image, attribute: .trailing, constant: 10.0)
-                    Anchors.bottom.equalToSuper(constant: -8.0)
-                    Anchors.trailing.equalToSuper(constant: -10.0)
-                }
-            }
-        }
-        
-        let expectedResult = """
-        ViewLayout - view: root
-        └─ ViewLayout - view: contentView
-           │     .leading == root.safeAreaLayoutGuide.leading + 16.0
-           │     .trailing == root.safeAreaLayoutGuide.trailing - 16.0
-           │     .centerY == root.centerY
-           │     .height == + 80.0
-           └─ TupleLayout
-              ├─ ViewLayout - view: image
-              │        .leading == superview.leading + 10.0
-              │        .centerY == superview.centerY
-              │        .width == + 70.0
-              │        .height == + 70.0
-              ├─ AnyLayout
-              │  └─ ViewLayout - view: title
-              │           .top == superview.top + 8.0
-              │           .leading == image.trailing + 10.0
-              │           .height == + 24.0
-              └─ ViewLayout - view: description
-                       .top == title.bottom + 5.0
-                       .leading == image.trailing + 10.0
-                       .bottom == superview.bottom - 8.0
-                       .trailing == superview.trailing - 10.0
-        """
-        
-        XCTAssertEqual(LayoutPrinter(layout, withAnchors: true).description, expectedResult)
-    }
 }
 
 extension ImplementationTests {
