+++ Tests/SwiftLayoutUtilTests/Util/Extensions.swift
@@ -1,0 +1,3 @@
+extension String {
+    var tabbed: String { replacingOccurrences(of: "    ", with: "\t") }
+}
